build(lint): Use npm-run-all to clean up composite scripts
This cleans up the pattern of `yarn run foo && yarn run bar ...` to
`run-{s,p} foo bar ...`, or even more compact with globs (as is done
here).  Normally a (seemingly) trivial dependency like this would be
avoided, but this also has a very nice feature of being able to run
scripts in parallel, and this reduces linting time by ~30% in initial
testing.

Also change to using `:` as script name delimiter (from `-`).  This
allows name components to be hyphenated without ambiguity and is more
similar to other tooling, e.g. rake.`